# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handler.UserResponseForSwagger:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      role:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  handler.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handler.updateUserRequest:
    properties:
      email:
        example: john.doe@example.com
        type: string
      full_name:
        example: John Doe
        type: string
      role:
        enum:
        - admin
        - instructor
        - student
        type: string
    type: object
...
  model.User:
    type: object
host: localhost:8080
info:
  contact:
    email: dimasrfebrian@gmail.com
    name: Dimas Rizky Febrian
    url: http://www.github.com/dimasrizkyfebrian
  description: This is the API documentation for the Coursify application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Coursify API
  version: "1.0"
paths:
  /admin/users/{id}:
    delete:
      description: Permanently deletes a user account.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a user (Admin only)
      tags:
      - Admin
    get:
      description: Retrieves the full details of a single user by their ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserResponseForSwagger'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a single user's details (Admin only)
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Updates a user's full_name, email, or role.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User data to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.updateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a user (Admin only)
      tags:
      - Admin
  /admin/users/{id}/approve:
    put:
      description: Changes a user's status from 'pending' to 'active'.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Approve a user (Admin only)
      tags:
      - Admin
  /admin/users/{id}/reject:
    put:
      description: Changes a user's status from 'pending' to 'rejected'.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reject a user (Admin only)
      tags:
      - Admin
  /admin/users/all:
    get:
      description: Retrieves a list of all users regardless of their status.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.UserResponseForSwagger'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all users (Admin only)
      tags:
      - Admin
  /admin/users/pending:
    get:
      description: Retrieves a list of users with 'pending' status.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.UserResponseForSwagger'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get pending users (Admin only)
      tags:
      - Admin
  /admin/users/pending/count:
    get:
      description: Retrieves the number of users with 'pending' status.
      produces:
      - application/json
      responses:
        "200":
          description: '{"count": 5}'
          schema:
            additionalProperties:
              type: integer
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get pending user count (Admin only)
      tags:
      - Admin
  /admin/users/stats:
    get:
      description: Retrieves key statistics like total, active, and pending users.
      produces:
      - application/json
      responses:
        "200":
          description: '{"total_users": 10, "active_users": 5, "pending_users": 2}'
          schema:
            additionalProperties:
              type: integer
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user statistics (Admin only)
      tags:
      - Admin
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token.
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"token": "JWT_TOKEN"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Log in a user
      tags:
      - Auth
  /profile:
    get:
      description: Retrieves the profile information for the currently logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserResponseForSwagger'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Users
  /profile/avatar:
    put:
      consumes:
      - multipart/form-data
      description: Uploads a new avatar image (jpg, png) for the logged-in user. Converts
        to WebP.
      parameters:
      - description: Avatar image file (jpg, png, max 2MB)
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: '{"message": "Avatar updated successfully", "url": "/uploads/avatars/..."}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: e.g., No file, file too large, invalid file type
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload or update user avatar
      tags:
      - Profile
  /register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with a 'pending' status.
      parameters:
      - description: User registration info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: '"Type ''Bearer'' followed by a space and a JWT token."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"